requirements.txt
fastapi==0.112.0
uvicorn==0.30.5
gunicorn==23.0.0
python-dotenv==1.0.1
pillow==10.4.0
tenacity==8.5.0
python-multipart==0.0.9
anthropic==0.40.0
pillow_heif==0.20.0
python-docx==1.1.2
reportlab==4.2.5